Audio refactor
